+++ kubeop-operator/apis/paas/v1alpha1/bucket_types.go
@@ -117,6 +117,7 @@ func (in *BucketStatus) DeepCopy() *BucketStatus {
 // +kubebuilder:printcolumn:name="AGE",type=date,JSONPath=`.metadata.creationTimestamp`
 // +kubebuilder:validation:XValidation:rule="has(self.metadata.labels['paas.kubeop.io/tenant'])",message="metadata.labels.paas.kubeop.io/tenant is required"
 // +kubebuilder:validation:XValidation:rule="has(self.metadata.labels['paas.kubeop.io/project'])",message="metadata.labels.paas.kubeop.io/project is required"
+// +kubebuilder:validation:XValidation:rule="has(self.metadata.labels['paas.kubeop.io/app'])",message="metadata.labels.paas.kubeop.io/app is required"
 // Bucket provisions object storage for workloads.
 type Bucket struct {
 	metav1.TypeMeta   `json:",inline"`